fix two typos in comments

+++ setup.py
@@ -18,13 +18,16 @@ Documentation with examples is available at http://packages.python.org/algopy/.
 
 # upload sphinx documentation
 #    python setup.py build_sphinx
-#    python setup.py upload_sphinx (pip install sphinx-pypi-upload to make this command available)
-# need to uncomment some code for that (look for comments containging build_sphinx)
+#    python setup.py upload_sphinx
+# need to uncomment some code for that (look for comments containing build_sphinx)
 
 DOCLINES = __doc__.split("\n")
 
 import os
+import shutil
 import sys
+import re
+import subprocess
 
 CLASSIFIERS = """\
 Intended Audience :: Science/Research
@@ -33,8 +36,8 @@ License :: OSI Approved
 Development Status :: 4 - Beta
 Operating System :: OS Independent
 Programming Language :: Python
+Programming Language :: Python
 Programming Language :: Python :: 3
-Topic :: Education
 Topic :: Software Development
 Topic :: Scientific/Engineering
 """
@@ -53,7 +56,7 @@ AUTHOR              = "Sebastian F. Walter"
 AUTHOR_EMAIL        = "sebastian.walter@gmail.com"
 PLATFORMS           = ["all"]
 MAJOR               = 0
-MINOR               = 5
+MINOR               = 4
 MICRO               = 0
 ISRELEASED          = False
 VERSION             = '%d.%d.%d' % (MAJOR, MINOR, MICRO)
@@ -86,13 +89,12 @@ if not release:
     version += '%(git_revision)s'
 """
 
+
     a = open(filename, 'w')
     try:
-        a.write(cnt % {'version': VERSION, 'isrelease': str(ISRELEASED),
-                       'git_revision': git_revision})
+        a.write(cnt % {'version': VERSION, 'isrelease': str(ISRELEASED), 'git_revision':git_revision})
     finally:
         a.close()
-
 
 def fullsplit(path, result=None):
     """
@@ -108,12 +110,10 @@ def fullsplit(path, result=None):
         return result
     return fullsplit(head, [tail] + result)
 
-
 def setup_package():
 
     # Rewrite the version file everytime
-    if os.path.exists('algopy/version.py'):
-        os.remove('algopy/version.py')
+    if os.path.exists('algopy/version.py'): os.remove('algopy/version.py')
     write_version_py()
 
     local_path = os.path.dirname(os.path.abspath(sys.argv[0]))
@@ -129,8 +129,7 @@ def setup_package():
     for dirpath, dirnames, filenames in os.walk('algopy'):
         # Ignore dirnames that start with '.'
         for i, dirname in enumerate(dirnames):
-            if dirname.startswith('.'):
-                del dirnames[i]
+            if dirname.startswith('.'): del dirnames[i]
         if '__init__.py' in filenames:
             packages.append('.'.join(fullsplit(dirpath)))
         elif filenames:
@@ -138,24 +137,24 @@ def setup_package():
 
     from distutils.core import setup
     # from distutils.core import setup, Extension
-    # from setuptools import setup  #uncomment for build_sphinx and upload_sphinx
+    # from setuptools import setup #uncomment for build_sphinx and upload_sphinx
 
     try:
         setup(name=NAME,
-              version=VERSION,
-              description=DESCRIPTION,
-              long_description=LONG_DESCRIPTION,
-              license=LICENSE,
-              author=AUTHOR,
-              platforms=PLATFORMS,
-              author_email=AUTHOR_EMAIL,
-              keywords=KEYWORDS,
-              url=URL,
-              packages=packages,
-              # ext_package='algopy.ctps',
-              # ext_modules=[Extension('libctps', ['algopy/ctps/src/ctps.c'])],
-              # entry_points = {"distutils.commands": ["upload_sphinx = sphinx_pypi_upload:UploadDoc",]} #uncomment for build_sphinx and upload_sphinx
-              )
+          version=VERSION,
+          description = DESCRIPTION,
+          long_description = LONG_DESCRIPTION,
+          license=LICENSE,
+          author=AUTHOR,
+          platforms=PLATFORMS,
+          author_email= AUTHOR_EMAIL,
+          keywords = KEYWORDS,
+          url=URL,
+          packages = packages,
+          # ext_package='algopy.ctps',
+          # ext_modules=[Extension('libctps', ['algopy/ctps/src/ctps.c'])],
+          # entry_points = {"distutils.commands": ["upload_sphinx = sphinx_pypi_upload:UploadDoc",]} #uncomment for build_sphinx and upload_sphinx
+         )
 
     finally:
         del sys.path[0]
